Quartz.ColorMap (quartz v0.1.0)

Functions

get_color(color_map, value)

Get color from a color_map for a given value.

pastel23()

Cyclic colormap with 8 colors.

From the D3.js project

Colors

pastel23(i)

Return a color for the given integer using the pastel23 colormap.

See pastel23/0.

set1()

Cyclic colormap with 9 colors.

From the D3.js project

Colors

set1(i)

Return a color for the given integer using the set1 colormap.

See set1/0.

set2()

Cyclic colormap with 8 colors.

From the D3.js project

Colors

set2(i)

Return a color for the given integer using the set2 colormap.

See set2/0.

set3()

Cyclic colormap with 12 colors.

From the D3.js project

Colors

set3(i)

Return a color for the given integer using the set3 colormap.

See set3/0.

tab10()

Cyclic colormap with 10 colors.

From the D3.js project

Colors

tab10(i)

Return a color for the given integer using the tab10 colormap.

See tab10/0.