Quartz.Config (Quartz v0.6.0)

Functions

axis_label_text_attributes(config, opts)

TODO: Document this.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from the config passed as an argument.

This function can be used outside a figure context.

default_config()

get_axis_label_text_attributes(opts \\ [])

Gets the default axis label text attributes.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

get_board_debug_properties()

TODO: Document this.

get_canvas_debug_properties()

TODO: Document this.

get_config()

TODO: Document this.

get_legend_text_attributes(opts \\ [])

Gets the default legend text attributes.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

get_major_tick_attributes(opts \\ [])

Gets the default major tick attributes.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

get_major_tick_label_text_attributes(opts \\ [])

Gets the default major tick label text attributes.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

get_major_tick_size(axis)

TODO: Document this.

get_math_character_attributes(opts \\ [])

Gets the default math character attributes.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

get_minor_tick_attributes(opts \\ [])

Gets the default minor tick attributes.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

get_plot_text_attributes(opts \\ [])

Gets the default plot text attributes.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

get_plot_title_text_attributes(opts \\ [])

Gets the default plot title attributes.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

get_text_debug_properties()

TODO: Document this.

legend_text_attributes(config, opts)

TODO: Document this.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from the config passed as an argument.

This function can be used outside a figure context.

major_tick_attributes(config, opts)

Extract major tick attributes from a config.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from the config passed as an argument.

This function can be used outside a figure context.

major_tick_label_text_attributes(config, opts)

TODO: Document this.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from the config passed as an argument.

This function can be used outside a figure context.

major_tick_size(config, opts)

TODO: Document this.

math_character_attributes(config, opts)

TODO: Document this.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from the config passed as an argument.

This function can be used outside a figure context.

minor_tick_attributes(config, opts)

xtract major tick attributes from a config.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from the config passed as an argument.

This function can be used outside a figure context.

new(opts \\ [])

TODO: Document this.

plot_text_attributes(config, opts)

TODO: Document this.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from the config passed as an argument.

This function can be used outside a figure context.

plot_title_text_attributes(config, opts)

TODO: Document this.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from the config passed as an argument.

This function can be used outside a figure context.

put_axis_label_text_attributes(item, opts \\ [])

Outs the the default axis label text attributes into the given item.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

put_legend_text_attributes(item, opts \\ [])

Outs the the default legend text attributes into the given item.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

put_major_tick_attributes(item, opts \\ [])

Outs the the default major tick attributes into the given item.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

put_major_tick_label_text_attributes(item, opts \\ [])

Outs the the default major tick label text attributes into the given item.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

put_math_character_attributes(item, opts \\ [])

Outs the the default math character attributes into the given item.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

put_minor_tick_attributes(item, opts \\ [])

Outs the the default minor tick attributes into the given item.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

put_new_attributes(item, attributes)

TODO: Document this.

put_plot_text_attributes(item, opts \\ [])

Outs the the default plot text attributes into the given item.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

put_plot_title_text_attributes(item, opts \\ [])

Outs the the default plot title attributes into the given item.

Optional arguments:

opts (default: []): a keyword list of options that override the default values taken from get_config/0

Only works inside a figure context.

serif_config()