Quartz.Figure (quartz v0.1.0)

Figure

Types

bounds()

@type bounds() :: Keyword.t()

t()

@type t() :: %Quartz.Figure{
  config: term(),
  counter: term(),
  debug: term(),
  finalized: term(),
  height: term(),
  plots: term(),
  problem: term(),
  resvg_options: term(),
  shared_axes: term(),
  sketches: term(),
  solution: term(),
  store: term(),
  unmeasured: term(),
  width: term()
}

Functions

add_plot_2d(fun)

add_sketch(id, sketch)

add_unmeasured_item(object)

Add an unmeasure item to the figure.

Unmeasured items are items in which the item dimensions can't be determined without rendering them first. Before rendering the final output, Quartz will render the unmeasured items and query the renderer for their dimensions.

assert(comparison)
(macro)

assert(left, operator, right)

assert_contains(container, item)

assert_contains_horizontally(container, item)

assert_contains_vertically(container, item)

assert_horizontally_contained_in(item, container)

assert_vertically_contained_in(item, container)

bounds_for_plots_in_grid(opts \\ [])

@spec bounds_for_plots_in_grid(Keyword.t()) :: [[bounds()]]

Get bounds for plots in a grid. Currently only regular grids are supported.

Supports the following options:

:nr_of_rows - required (integer)

:nr_of_columns - required (integer)

:padding - optional (default: Length.pt(8)): the padding between plots

:horizontal_padding - optional (default: the value of :padding): the horizontal padding between plots

:vertical_padding - optional (default: the value of :padding): the vertical padding between plots

Returns a neste list of bounds.

center_on(object, container)

current_figure_height()

current_figure_width()

debug?()

Returns true if the current figure is being created with :debug mode active.

get_current_figure()

get_id()

maximize(expression, opts \\ [])
(macro)

minimize(expression, opts \\ [])
(macro)

new(args, fun)

Create a new figure.

place_in_canvas(sketch, canvas, opts)

Place a sketch in a canvas with the given parameters.s

position_with_location_and_alignment(sketch, container, opts \\ [])

TODO: document this

put_current_figure(figure)

put_plot_in_current_figure(plot)

render_to_png_binary(figure)

@spec render_to_png_binary(Figure.t()) :: String.t()

Renders a figure into a PNG binary.

render_to_png_file(figure, path)

@spec render_to_png_file(Figure.t(), Path.t()) :: :ok | {:error, binary()}

Renders a figure into a PNG file and saves it in the given path.

render_to_png_file!(figure, path)

@spec render_to_png_file!(Figure.t(), Path.t()) :: :ok | {:error, binary()}

Renders a figure into a PNG file and saves it in the given path. Raises an error if it can't write to the file.

render_to_svg_binary(figure)

@spec render_to_svg_binary(Figure.t()) :: String.t()

Renders a figure into an SVG string (not iodata).

render_to_svg_file(figure, path)

@spec render_to_svg_file(Figure.t(), Path.t()) :: :ok | {:error, File.posix()}

Renders a figure into SVG and saves it in the given path.

render_to_svg_file!(figure, path)

@spec render_to_svg_file!(Figure.t(), Path.t()) :: :ok

Renders a figure into SVG and saves it in the given path. Raises an error if it can't write to the file.

render_to_svg_iodata(figure)

Renders a figure into SVG iodata (not a string).

share_axes(axes)

solve!(expression)

solve_sketch!(sketch)

stack_horizontally(objects)

stack_horizontally(objects, list)

stack_vertically(objects)

stack_vertically(objects, list)

update_current_figure(fun)

variable(name, opts \\ [])