Quartz.Formatter (Quartz v0.6.0)

Functions to help display floats and dates.

Functions

rounded_float(float, nr_of_decimal_places, format \\ :normal)

Display a float rounded to the given number of decimals.

rounded_length(float, decimals \\ 5)

A function to display rounded length with a given number of decimals.

This is used mainly when rendering sketches to SVG. The goal of defining this function here is to centralize the default number of decimal places for floats that are used as part of drawing dimensions.

This function is public API and is meant to be used by end users who might want to implement new sketches.