Quartz.LegendSymbolBuilder (Quartz v0.6.0)

TODO

Functions

%Quartz.LegendSymbolBuilder{}
(struct)

A symbol builder.

build(symbol, label, builder_opts)

Builds a symbol from a symbol builder and a concrete legend.

circle(opts)

Creates a circle symbol builder for a legend.

get_label_height(label)

Get the height of a label to draw a symbol marker.

line(opts)

Creates a line symbol builder for a legend.

rectangle(opts)

Creates a rectangle symbol builder for a legend.