Quartz.SVG (quartz v0.1.0)

Utilities to work with SVG elements.

Types

t()

@type t() :: {binary(), list(), list()}

An SVG element is represented by

Functions

attr_value_to_iodata(color)

attrs_to_iodata(attrs)

circle(attrs, content \\ [])

SVG <circle/> element.

escaped_iodata(iodata)

g(attrs, contents)

SVG <g/> element.

line(attrs, content \\ [])

SVG <line/> element.

path(attrs, content \\ [])

Build a path.

rect(attrs, content \\ [])

SVG <rect/> element.

style_to_iodata(style_attrs)

svg(attrs, contents)

Build a top level SVG element.

text(attrs, content)

SVG <text/> element.

title(attrs, content)

SVG <title/> element.

This element is not rendered but is very useful for debugging. It can be used to show properties of the containing element.

to_binary(element)

to_iodata(list_of_stuff)

tspan(attrs, content)

SVG <tspan/> element.

xml_escape(data)

@spec xml_escape(any()) :: any()

xml_escape_iodata(data)

@spec xml_escape_iodata(iodata()) :: {:escaped_iodata, iodata()}

xml_escape_to_iodata(data)

@spec xml_escape_to_iodata(String.t()) :: {:escaped_iodata, iodata()}

Escapes the given HTML to iodata.

iex> Quartz.SVG.xml_escape_to_iodata("foo")
{:escaped_iodata, "foo"}

iex> Quartz.SVG.xml_escape_to_iodata("<foo>")
{:escaped_iodata, [[[] | "&lt;"], "foo" | "&gt;"]}

iex> Quartz.SVG.xml_escape_to_iodata("quotes: \" & \'")
{:escaped_iodata, [[[[], "quotes: " | "&quot;"], " " | "&amp;"], " " | "&#39;"]}